add bank account

+++ repository/bank_account._repository.go
@@ -9,20 +9,14 @@ import (
 
 type BankAccRepository interface {
 	GetAll() any
-	GetByID(id int) any
+	GetByID(id uint) any
 	Create(newBankAcc *model.BankAcc) string
-	Update(bankacc *model.BankAcc) string
-	Delete(id int) string
+	Update(bankAcc *model.BankAcc) string
+	Delete(id uint) string
 }
 
 type bankAccRepository struct {
 	db *sql.DB
-}
-
-func NewBankAccRepository(db *sql.DB) BankAccRepository {
-	repo := new(bankAccRepository)
-	repo.db = db
-	return repo
 }
 
 func (r *bankAccRepository) GetAll() any {
@@ -53,12 +47,12 @@ func (r *bankAccRepository) GetAll() any {
 	return users
 }
 
-func (r *bankAccRepository) GetByID(id int) any {
+func (r *bankAccRepository) GetByID(id uint) any {
 	var user model.BankAcc
 
-	query := "SELECT account_id, bank_account, account_number, account_holder_name WHERE user_id = $1"
+	query := "SELECT bank_name, account_number, account_holder_name FROM mst_bank_account WHERE user_id = $1"
 	row := r.db.QueryRow(query, id)
-	err := row.Scan(&user.AccountId, &user.BankName, &user.AccountNumber, &user.AccountHolderName)
+	err := row.Scan(&user.BankName, &user.AccountNumber, &user.AccountHolderName)
 	if err != nil {
 		log.Println(err)
 	}
@@ -66,8 +60,8 @@ func (r *bankAccRepository) GetByID(id int) any {
 	return user
 }
 func (r *bankAccRepository) Create(newBankAcc *model.BankAcc) string {
-	query := "INSERT INTO mst_bank_account (account_id, user_id, bank_account, account_number, account_holder_name) VALUES ($1, $2, $3, $4, $5)"
-	_, err := r.db.Exec(query, newBankAcc.AccountId, newBankAcc.UserId, newBankAcc.BankName, newBankAcc.AccountNumber, newBankAcc.AccountHolderName)
+	query := "INSERT INTO mst_bank_account (user_id, bank_name, account_number, account_holder_name) VALUES ($1, $2, $3, $4)"
+	_, err := r.db.Exec(query, newBankAcc.UserId, newBankAcc.BankName, newBankAcc.AccountNumber, newBankAcc.AccountHolderName)
 	if err != nil {
 		log.Println(err)
 		return "failed to create data"
@@ -77,14 +71,14 @@ func (r *bankAccRepository) Create(newBankAcc *model.BankAcc) string {
 }
 
 func (r *bankAccRepository) Update(bankAcc *model.BankAcc) string {
-	result := r.GetByID(bankAcc.UserId)
+	result := r.GetByID(bankAcc.AccountId)
 
 	if result == "Bank Account not found" {
 		return result.(string)
 	}
 
-	query := "UPDATE mst_bank_account SET account_id = $1, bank_account = $2, account_number = $3, account_holder_name = $4 WHERE user_id = $5"
-	_, err := r.db.Exec(query, bankAcc.AccountId, bankAcc.BankName, bankAcc.AccountNumber, bankAcc.AccountHolderName)
+	query := "UPDATE mst_bank_account SET user_id = $1, bank_name = $2, account_number = $3, account_holder_name = $4 WHERE user_id = $5"
+	_, err := r.db.Exec(query, bankAcc.UserId, bankAcc.BankName, bankAcc.AccountNumber, bankAcc.AccountHolderName)
 	if err != nil {
 		log.Println(err)
 		return "failed to update Bank Account"
@@ -93,7 +87,13 @@ func (r *bankAccRepository) Update(bankAcc *model.BankAcc) string {
 	return "Bank Account updated Successfully"
 }
 
-func (r *bankAccRepository) Delete(id int) string {
+func (r *bankAccRepository) Delete(id uint) string {
+	result := r.GetByID(id)
+
+	if result == "Bank Account not found" {
+		return result.(string)
+	}
+
 	query := "DELETE FROM mst_bank_account WHERE user_id = $1"
 	_, err := r.db.Exec(query, id)
 	if err != nil {
@@ -102,4 +102,10 @@ func (r *bankAccRepository) Delete(id int) string {
 	}
 
 	return "Bank Account deleted Successfully"
+}
+
+func NewBankAccRepository(db *sql.DB) BankAccRepository {
+	repo := new(bankAccRepository)
+	repo.db = db
+	return repo
 }
